Moved TTYGService to js/angular/core/service module
